Logging System Created
Created a logging system
This system works just like std::cout and it will output to the terminal. However, it also writes everything to a buffer which is then outputted to file after a successful execution.
A function for capturing the time was created so as to date every log file and every message where relevant.
The only part of the code that should still use cout is the part that says whether the log file was saved properly.
Changed the gitignore so that logging files are not committed